Fix test issues due to different execution order in Gradle
